Remove mowgli dependency, deprecate libguess_init()
Using a patricia tree and pulling in another library for it seems
like overkill. In particular, initializing the tree was not done
in a safe way, and required the user to call libguess_init().

Replace this with a simple static table. It does linear search, but
this likely will not matter for 11 entries. libguess_init() is now
an empty stub, left only so that programs using will still compile.